requirements.txt
torch>=2.0.0
torchvision>=0.15.0
streamlit>=1.32.0
Pillow>=9.0.0
numpy>=1.23.0
